% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametricboot-package.R
\docType{package}
\name{parametricboot-package}
\alias{parametricboot}
\alias{parametricboot-package}
\title{parametricboot: Parametric Bootstrap for Fitted Regression Models}
\description{
Automates the parametric bootstrap for models fitted with \code{\link[stats:lm]{stats::lm()}},
\code{\link[stats:glm]{stats::glm()}}, \code{\link[lme4:lmer]{lme4::lmer()}}, \code{\link[lme4:glmer]{lme4::glmer()}} and \code{\link[survival:coxph]{survival::coxph()}}.
}
\section{Workflow}{

\enumerate{
\item Fit a model as usual.
\item Draw bootstrap replicates with \code{\link[=pb_simulate]{pb_simulate()}}, or \code{\link[=pb_parallel]{pb_parallel()}} when
a single refit is slow.
\item Summarise them: \code{\link[=pb_summary_table]{pb_summary_table()}}, \code{\link[=pb_key_stats]{pb_key_stats()}}, \code{\link[=pb_confint]{pb_confint()}}
and \code{\link[=pb_tidy]{pb_tidy()}} for tables; \code{\link[=pb_plot_estimates]{pb_plot_estimates()}} and
\code{\link[=pb_plot_diagnostics]{pb_plot_diagnostics()}} for plots. If some refits misbehaved, see
\code{\link[=pb_drop_warned]{pb_drop_warned()}}.
\item Propagate the uncertainty to predictions with \code{\link[=pb_predict_boot]{pb_predict_boot()}} and
\code{\link[=pb_plot_predictions]{pb_plot_predictions()}}, or to model comparisons with
\code{\link[=pb_compare_models]{pb_compare_models()}}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://diogoribeiro7.github.io/parametricboot/}
  \item \url{https://github.com/DiogoRibeiro7/parametricboot}
  \item Report bugs at \url{https://github.com/DiogoRibeiro7/parametricboot/issues}
}

}
\author{
\strong{Maintainer}: Diogo Ribeiro \email{diogo.debastos.ribeiro@gmail.com} (\href{https://orcid.org/0009-0001-2022-7072}{ORCID}) (ESMAD, Instituto Politecnico do Porto) [copyright holder]

}
\keyword{internal}
