\name{pb_key_stats}
\alias{pb_key_stats}
\title{Compute bias, MSE and coverage}
\description{Calculate key statistics from bootstrap results.}
\usage{pb_key_stats(results)}
\arguments{
  \item{results}{Output from \code{pb_simulate()}.}
}
\value{A data frame with bias, MSE and coverage.}
\examples{
## pb_key_stats(res)
}
