\name{pb_resample}
\alias{pb_resample}
\title{Resample a data frame with replacement}
\description{Randomly sample rows from a data frame.}
\usage{pb_resample(data, size = nrow(data))}
\arguments{
  \item{data}{Data frame to resample.}
  \item{size}{Number of rows to sample.}
}
\value{A resampled data frame.}
\examples{
## pb_resample(df)
}
